Implemented a basic Tic Tac Toe game in Python (User-friendly console interface)
Tic tac toe Python, also known as Noughts and Crosses or Xs and Os, is a very simple two-player game where both the player choose between the symbol 'X' and 'O' to mark the position on a 3X3 grid. Once a player is successful in marking a strike of the same symbol either in the horizontal, vertical, or diagonal way that player wins the game else the game goes on a draw if all the spots are filled.

I had created the game with the same logic by using the fundamental concepts of python. I had the divided this mini-project program into different steps and then solved each step one by one.
The game that I had created is the implementation of Tic Tac Toe game but this is not a GUI interfaced game, it is a CLI interfaced.
Thank You!